requirements.txt
PyYAML~=6.0
requests~=2.28.1
beautifulsoup4~=4.11.1
dateparser~=1.1.1